requirements.txt
ollama
chromadb
PyPDF2
python-docx
